Add screenshot test for tutorial page
